+++ src/utils/clients/discogs/methods/albumGetInfo.ts
@@ -1,6 +1,6 @@
-import { discogsAPI } from '../api-client';
-import { albumGetInfoTransformer } from '../transformers/album-get-info-response.transformer';
-import { tracksTransformer } from '../transformers/tracks-response.transformer';
+import { discogsAPI } from '../apiClient';
+import { albumGetInfoTransformer } from '../transformers/albumGetInfoResponse.transformer';
+import { tracksTransformer } from '../transformers/tracksResponse.transformer';
 
 export async function albumGetInfo(discogsId: string) {
   const response = await discogsAPI.get('', {
