+++ src/utils/clients/discogs/methods/searchTopAlbums.test.ts
@@ -1,7 +1,7 @@
-import { discogsAPI } from '../api-client';
-import { searchTopAlbums } from './search-top-albums';
+import { discogsAPI } from '../apiClient';
+import { searchTopAlbums } from './searchTopAlbums';
 
-jest.mock('../api-client');
+jest.mock('../apiClient');
 
 describe('the `searchTopAlbums` function', () => {
   it('should call the API with the given discogs id', async() => {
