+++ src/utils/clients/discogs/methods/searchTopAlbums.ts
@@ -1,5 +1,5 @@
-import { discogsAPI } from '../api-client';
-import { topAlbumsTransformer } from '../transformers/top-albums-response.transformer';
+import { discogsAPI } from '../apiClient';
+import { topAlbumsTransformer } from '../transformers/topAlbumsResponse.transformer';
 
 export async function searchTopAlbums(discogsId: string) {
   const response = await discogsAPI.get('', {
