+++ src/utils/clients/discogs/transformers/albumGetInfoResponse.transformer.test.ts
@@ -1,4 +1,4 @@
-import { albumGetInfoTransformer } from './album-get-info-response.transformer';
+import { albumGetInfoTransformer } from './albumGetInfoResponse.transformer';
 
 describe('the `albumGetInfoTransformer` function', () => {
   it('should return an empty object if there was a problem retrieving the name', () => {
