+++ src/utils/clients/discogs/transformers/common/sanitizeArtistName.ts
