+++ src/utils/clients/lastfm/methods/albumGetInfo.ts
@@ -1,6 +1,6 @@
-import { lastfmAPI } from '../api-client';
-import { albumGetInfoTransformer } from '../transformers/album-get-info-response.transformer';
-import { tracksTransformer } from '../transformers/tracks-response.transformer';
+import { lastfmAPI } from '../apiClient';
+import { albumGetInfoTransformer } from '../transformers/albumGetInfoResponse.transformer';
+import { tracksTransformer } from '../transformers/tracksResponse.transformer';
 
 export async function albumGetInfo(album: { mbid?: string; artist: string; name: string }) {
   const searchParams = album.mbid
